requirements.txt
python-telegram-bot[job-queue]==21.3
groq
aiohttp
feedparser
python-dotenv
